backend/requirements.txt
python-dotenv==1.0.0
Flask==2.3.3
Flask-Cors==4.0.0
ultralytics==8.0.159